+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: CKV_AWS_326: "Ensure that RDS Aurora Clusters have backtracking enabled"
 - Refactor cluster instances block to allow for unique configurations per instance.(use for each statement)
 - Restore from s3 bucket example
-- parameter group example for cluster instances
+- feat: parameter group for cluster instances
 - Example IAM roles attached to cluster
 - Move values to variables.tf files in examples